update 'first' example
