+++ example/first/Cargo.toml
@@ -3,8 +3,12 @@ name = "first"
 version = "0.1.0"
 authors = ["Jun Kimura <junkxdev@gmail.com>"]
 edition = "2018"
-build = "build.rs"
 
 [dependencies]
 hmc = { git = "https://github.com/bluele/hypermint", tag = "v0.3.0" }
+
+[dev-dependencies]
 hmemu = { path = "../.." }
+
+[lib]
+crate-type = ["cdylib"]
