--- example/first/build.rs
@@ -1,5 +1,0 @@
-
-fn main(){
-    println!(r"cargo:rustc-link-lib=dylib=hm");
-    println!(r"cargo:rustc-link-search=native=../../build");
-}
